+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Added components to visualise canopy/root masses using meshes
 - Flow components can now trace uphill paths with negative distances
 - Flow catchments component now guesses a proximity threshold if one is not set
+- Plant components actually use variance values
 
 #### [0.10.2] - 2019-04-09
 ###### Fixed